Assemble focus trap helper
Dialogs and drawers should trap focus when navigation is done using a keyboard in order to prevent users from accidentally winding up somewhere else on the page. Other items may require this functionality too so it is best to abstract it as a helper.
