+++ source/components/dialog.js
@@ -3,6 +3,7 @@
 
 import manageTransit from "../helpers/manage-transit";
 import toggleScroll from "../helpers/toggle-scroll";
+import trapFocus from "../helpers/trap-focus";
 
 const SELECTOR_DIALOG = ".dialog";
 const SELECTOR_BACKDROP = ".dialog-backdrop";
@@ -70,17 +71,7 @@ function Dialog(dialog) {
 
   function handleFocusTrap(event) {
     if (event.key === "Tab") {
-      const focusableElements = Array.from(dialog.querySelectorAll("a[href], audio[controls], button:not([disabled]), details, input:not([disabled]), select:not([disabled]), textarea:not([disabled]), video[controls], [contenteditable]"));
-      const lastIndex = focusableElements.length - 1;
-      const focusIndex = focusableElements.indexOf(document.activeElement);
-
-      if ((event.shiftKey && focusIndex === 0) || (event.shiftKey && document.activeElement === dialog)) {
-        event.preventDefault();
-        focusableElements[focusableElements.length - 1].focus();
-      } else if (!event.shiftKey && focusIndex === lastIndex) {
-        event.preventDefault();
-        focusableElements[0].focus();
-      }
+      trapFocus(event, dialog);
     }
   }
